Changed mmPerRev from int to float.
git-svn-id: http://polargraph.googlecode.com/svn/embedded/trunk@249 5ebda15a-75ab-8a5e-e99a-b4cc1c219dce

+++ polargraph_server_a1.ino
@@ -676,7 +676,7 @@ void setMachineNameFromCommand()
 
 void setMachineMmPerRevFromCommand()
 {
-  int mmPerRev = asInt(inParam1);
+  float mmPerRev = asFloat(inParam1);
   EEPROMWriteInt(EEPROM_MACHINE_MM_PER_REV, mmPerRev);
   loadMachineSpecFromEeprom();
 }
